Implement search query & displaying of results
Create new attachable part, `SearchablePart`, which can be added to a
content type to make that content type included within the search query.
The part has a couple of configuration options. One is the ability to
exclude the item from search results. The other is a textarea for
defining additional terms that if matched will include the content
within the search results.

The site search content item now includes the ability to define the
page size for the search results.