Refactor to use Queries instead of database query via YesSql
Refactored method of getting search results to use `OrchardCore.Queries`
instead of a fixed database query constructed via `ISession`. This
change opens the possibility of a deeper search that can attempt to
match the content with the content item instead of the title &
configurable keywords.

"SiteSearch" now has a `Query` field that contains all the queries in
the CMS. When using "Grouped" to display the results, each content type
that has the `SearchablePart` can have it's own "Query" configured.